% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint.R
\name{getcarcass_sol_fx_trait}
\alias{getcarcass_sol_fx_trait}
\title{Get for an effect key/code the solutions}
\usage{
getcarcass_sol_fx_trait(
  ptbl_sol_fix,
  ptbl_level_code,
  psTrait,
  psEffectnameInSolani
)
}
\arguments{
\item{ptbl_sol_fix}{tibble with solution of the fixed effect}

\item{ptbl_level_code}{tiblle with code and key of the fixed effect}

\item{psTrait}{name of the trait}

\item{psEffectnameInSolani}{name of the fixed effect in the Solani-file}
}
\value{
tbl_sol_fx_trt with solution as well as code and key of the fixed effects
}
\description{
Read csv-file generate from EBV-carcass-process
countaining the effect key and code.
}
