% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read_level_code}
\alias{read_level_code}
\title{Read CSV-file generate in carcass-process with effect key and code}
\usage{
read_level_code(pvec_path)
}
\arguments{
\item{pvec_path}{vector of path to csv-file}
}
\value{
tbl_level_code with code and key of the fixed effects
}
\description{
Read csv-file generated from EBV-carcass-process
countaining the effect key and code.
}
