% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read_solfix}
\alias{read_solfix}
\title{Read Solfix File of MiX99}
\usage{
read_solfix(pvec_dir)
}
\arguments{
\item{pvec_dir}{vector of directory to Solfix}
}
\value{
tbl_sol_fix with solution of the fixed effects
}
\description{
Read Solfix file generated with MiX99 where the solution
of the different fixed effect are in this file.
}
